refactor: rename src/tools to src/safeoutputs, create src/tools for first-class tools
Rename the existing tools directory to safeoutputs to better reflect its
purpose (safe-output MCP tool implementations that serialize to NDJSON
in Stage 1 and execute in Stage 2).

Create a new src/tools directory for first-class tool implementations
that the compiler auto-configures (cache-memory, azure-devops).

Move memory.rs from safeoutputs to tools/cache_memory.rs since memory
is a first-class tool, not a safe-output.

Add CacheMemoryToolConfig and AzureDevOpsToolConfig types to
compile/types.rs with support for both boolean and object front-matter
formats. Extend ToolsConfig to include cache-memory and azure-devops
fields.
