+++ src/safeoutputs/create_git_tag.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 use super::{PATH_SEGMENT, validate_git_ref_name};
 use crate::sanitize::{Sanitize, sanitize as sanitize_text};
 use crate::tool_result;
-use crate::tools::{ExecutionContext, ExecutionResult, Executor, Validate};
+use crate::safeoutputs::{ExecutionContext, ExecutionResult, Executor, Validate};
 use anyhow::{Context, ensure};
 
 /// Parameters for creating a git tag (agent-provided)
@@ -392,7 +392,7 @@ impl Executor for CreateGitTagResult {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::ToolResult;
+    use crate::safeoutputs::ToolResult;
 
     #[test]
     fn test_result_has_correct_name() {
