+++ src/safeoutputs/create_wiki_page.rs
@@ -1,4 +1,4 @@
-//! Update wiki page safe output tool
+//! Create wiki page safe output tool
 
 use anyhow::{Context, ensure};
 use log::{debug, info};
@@ -10,24 +10,24 @@ use super::PATH_SEGMENT;
 use super::resolve_wiki_branch;
 use crate::sanitize::{Sanitize, sanitize as sanitize_text};
 use crate::tool_result;
-use crate::tools::{ExecutionContext, ExecutionResult, Executor, Validate};
+use crate::safeoutputs::{ExecutionContext, ExecutionResult, Executor, Validate};
 
-/// Parameters for editing a wiki page (agent-provided)
+/// Parameters for creating a wiki page (agent-provided)
 #[derive(Deserialize, JsonSchema)]
-pub struct UpdateWikiPageParams {
-    /// Path of the wiki page to update, e.g. "/Overview/Architecture".
-    /// The page must already exist. The path must not contain "..".
+pub struct CreateWikiPageParams {
+    /// Path of the wiki page to create, e.g. "/Overview/NewPage".
+    /// The page must not already exist. The path must not contain "..".
     pub path: String,
 
     /// Markdown content for the wiki page. Must be at least 10 characters.
     pub content: String,
 
     /// Optional commit comment describing the change. Defaults to the value
-    /// configured in the front matter (or "Updated by agent" if not set).
+    /// configured in the front matter (or "Created by agent" if not set).
     pub comment: Option<String>,
 }
 
-impl Validate for UpdateWikiPageParams {
+impl Validate for CreateWikiPageParams {
     fn validate(&self) -> anyhow::Result<()> {
         ensure!(!self.path.trim().is_empty(), "path must not be empty");
         ensure!(
@@ -38,6 +38,10 @@ impl Validate for UpdateWikiPageParams {
             !self.path.contains(".."),
             "path must not contain '..': {}",
             self.path
+        );
+        ensure!(
+            self.path.trim_matches('/') != "",
+            "path must contain at least one non-slash segment"
         );
         ensure!(
             !self.content.is_empty(),
@@ -52,18 +56,18 @@ impl Validate for UpdateWikiPageParams {
 }
 
 tool_result! {
-    name = "update-wiki-page",
+    name = "create-wiki-page",
     write = true,
-    params = UpdateWikiPageParams,
-    /// Result of editing a wiki page
-    pub struct UpdateWikiPageResult {
+    params = CreateWikiPageParams,
+    /// Result of creating a wiki page
+    pub struct CreateWikiPageResult {
         path: String,
         content: String,
         comment: Option<String>,
     }
 }
 
-impl Sanitize for UpdateWikiPageResult {
+impl Sanitize for CreateWikiPageResult {
     fn sanitize_fields(&mut self) {
         // Path is a structural identifier — sanitize lightly (remove control chars)
         // but do not escape HTML or neutralize patterns that are valid in wiki paths.
@@ -81,19 +85,19 @@ impl Sanitize for UpdateWikiPageResult {
 // Front-matter configuration
 // ============================================================================
 
-/// Configuration for the `update-wiki-page` tool (specified in front matter).
+/// Configuration for the `create-wiki-page` tool (specified in front matter).
 ///
 /// ```yaml
 /// safe-outputs:
-///   update-wiki-page:
+///   create-wiki-page:
 ///     wiki-name: "MyProject.wiki"
 ///     wiki-project: "OtherProject"  # optional, defaults to current project
 ///     path-prefix: "/agent-output"
 ///     title-prefix: "[Agent] "
-///     comment: "Updated by agent"
+///     comment: "Created by agent"
 /// ```
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct UpdateWikiPageConfig {
+pub struct CreateWikiPageConfig {
     /// Wiki identifier (name or ID). Required — execution fails without this.
     ///
     /// For a project wiki, the identifier is typically `<ProjectName>.wiki`.
@@ -112,7 +116,7 @@ pub struct UpdateWikiPageConfig {
     #[serde(default)]
     pub branch: Option<String>,
 
-    /// Security restriction: the agent may only write wiki pages whose paths
+    /// Security restriction: the agent may only create wiki pages whose paths
     /// start with this prefix (e.g. `"/agent-output"`). Paths that do not match
     /// are rejected at execution time. When omitted, no restriction is applied.
     #[serde(default, rename = "path-prefix")]
@@ -164,9 +168,9 @@ fn apply_title_prefix(path: &str, prefix: &str) -> String {
 // ============================================================================
 
 #[async_trait::async_trait]
-impl Executor for UpdateWikiPageResult {
+impl Executor for CreateWikiPageResult {
     async fn execute_impl(&self, ctx: &ExecutionContext) -> anyhow::Result<ExecutionResult> {
-        info!("Editing wiki page: '{}'", self.path);
+        info!("Creating wiki page: '{}'", self.path);
         debug!("Content length: {} chars", self.content.len());
 
         let org_url = ctx
@@ -182,12 +186,12 @@ impl Executor for UpdateWikiPageResult {
             .as_ref()
             .context("No access token available (SYSTEM_ACCESSTOKEN or AZURE_DEVOPS_EXT_PAT)")?;
 
-        let config: UpdateWikiPageConfig = ctx.get_tool_config("update-wiki-page");
+        let config: CreateWikiPageConfig = ctx.get_tool_config("create-wiki-page");
 
         let wiki_name = config
             .wiki_name
             .as_deref()
-            .context("wiki-name must be configured in safe-outputs.update-wiki-page.wiki-name")?;
+            .context("wiki-name must be configured in safe-outputs.create-wiki-page.wiki-name")?;
 
         // Use the wiki-project override if present, otherwise use the pipeline project.
         let project = config
@@ -251,7 +255,7 @@ impl Executor for UpdateWikiPageResult {
             Err(msg) => return Ok(ExecutionResult::failure(msg)),
         };
 
-        // ── GET: check whether the page exists and obtain its ETag ────────────
+        // ── GET: check whether the page already exists ────────────────────────
         let mut get_query: Vec<(&str, &str)> = vec![
             ("path", effective_path.as_str()),
             ("api-version", "7.0"),
@@ -279,28 +283,28 @@ impl Executor for UpdateWikiPageResult {
 
         let page_exists = get_status.is_success();
 
-        if !page_exists {
+        if page_exists {
             return Ok(ExecutionResult::failure(format!(
-                "Wiki page '{effective_path}' does not exist. \
-                 Use a separate safe output to create new pages."
+                "Wiki page '{effective_path}' already exists. \
+                 Use the update-wiki-page safe output to update existing pages."
             )));
         }
-
-        let etag: Option<String> = get_resp
-            .headers()
-            .get("ETag")
-            .and_then(|v| v.to_str().ok())
-            .map(|s| s.to_string());
 
         let comment = self
             .comment
             .as_deref()
             .or(config.comment.as_deref())
-            .unwrap_or("Updated by agent");
+            .unwrap_or("Created by agent");
 
-        debug!("Updating existing wiki page: {effective_path}");
+        debug!("Creating new wiki page: {effective_path}");
 
-        // ── PUT: create or update the page ────────────────────────────────────
+        // ── PUT: create the new page using If-Match: "" for atomic create-only ──
+        //
+        // The ADO Wiki Pages API treats a PUT without If-Match as an upsert.
+        // Sending `If-Match: ""` (empty ETag) tells the API "create only — fail
+        // with 412 if this resource already exists", closing the TOCTOU race
+        // between our GET (404) and the PUT where a concurrent request could
+        // create the page first.
         let mut put_query: Vec<(&str, &str)> = vec![
             ("path", effective_path.as_str()),
             ("comment", comment),
@@ -310,24 +314,29 @@ impl Executor for UpdateWikiPageResult {
             put_query.push(("versionDescriptor.version", branch.as_str()));
             put_query.push(("versionDescriptor.versionType", "branch"));
         }
-        let mut put_req = client
+        let put_resp = client
             .put(&base_url)
             .query(&put_query)
             .header("Content-Type", "application/json")
+            .header("If-Match", "")
             .basic_auth("", Some(token))
-            .json(&serde_json::json!({ "content": self.content }));
-
-        // Provide the ETag for optimistic concurrency when updating an existing page.
-        if let Some(etag) = &etag {
-            put_req = put_req.header("If-Match", etag);
-        }
-
-        let put_resp = put_req
+            .json(&serde_json::json!({ "content": self.content }))
             .send()
             .await
-            .context("Failed to write wiki page")?;
+            .context("Failed to create wiki page")?;
 
-        if put_resp.status().is_success() {
+        let put_status = put_resp.status();
+
+        // 412 Precondition Failed means the page was created between our GET
+        // and our PUT (TOCTOU). Surface this as a clean, actionable message.
+        if put_status.as_u16() == 412 {
+            return Ok(ExecutionResult::failure(format!(
+                "Wiki page '{effective_path}' already exists (conflict detected during creation). \
+                 Use the update-wiki-page safe output to update existing pages."
+            )));
+        }
+
+        if put_status.is_success() {
             let body: serde_json::Value = put_resp.json().await.unwrap_or_default();
             let page_id = body
                 .get("id")
@@ -339,24 +348,23 @@ impl Executor for UpdateWikiPageResult {
                 .and_then(|v| v.as_str())
                 .unwrap_or("");
 
-            info!("Updated wiki page: {effective_path} (id={page_id})");
+            info!("Created wiki page: {effective_path} (id={page_id})");
 
             Ok(ExecutionResult::success_with_data(
-                format!("Updated wiki page: {effective_path}"),
+                format!("Created wiki page: {effective_path}"),
                 serde_json::json!({
                     "id": page_id,
                     "path": effective_path,
                     "url": remote_url,
                     "wiki": wiki_name,
                     "project": project,
-                    "action": "updated",
+                    "action": "created",
                 }),
             ))
         } else {
-            let put_status = put_resp.status();
             let error_body = put_resp.text().await.unwrap_or_default();
             Ok(ExecutionResult::failure(format!(
-                "Failed to update wiki page '{}' (HTTP {}): {}",
+                "Failed to create wiki page '{}' (HTTP {}): {}",
                 effective_path,
                 put_status,
                 error_body
@@ -372,19 +380,19 @@ impl Executor for UpdateWikiPageResult {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::ToolResult;
+    use crate::safeoutputs::ToolResult;
 
     // ── ToolResult / macro ────────────────────────────────────────────────────
 
     #[test]
     fn test_result_has_correct_name() {
-        assert_eq!(UpdateWikiPageResult::NAME, "update-wiki-page");
+        assert_eq!(CreateWikiPageResult::NAME, "create-wiki-page");
     }
 
     #[test]
     fn test_params_deserializes() {
         let json = r#"{"path": "/Overview", "content": "Hello, wiki!"}"#;
-        let params: UpdateWikiPageParams = serde_json::from_str(json).unwrap();
+        let params: CreateWikiPageParams = serde_json::from_str(json).unwrap();
         assert_eq!(params.path, "/Overview");
         assert_eq!(params.content, "Hello, wiki!");
         assert!(params.comment.is_none());
@@ -393,19 +401,19 @@ mod tests {
     #[test]
     fn test_params_with_comment_deserializes() {
         let json = r#"{"path": "/Overview", "content": "Hello, wiki!", "comment": "initial"}"#;
-        let params: UpdateWikiPageParams = serde_json::from_str(json).unwrap();
+        let params: CreateWikiPageParams = serde_json::from_str(json).unwrap();
         assert_eq!(params.comment, Some("initial".to_string()));
     }
 
     #[test]
     fn test_params_converts_to_result() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/My Page".to_string(),
             content: "Some wiki content here".to_string(),
             comment: None,
         };
-        let result: UpdateWikiPageResult = params.try_into().unwrap();
-        assert_eq!(result.name, "update-wiki-page");
+        let result: CreateWikiPageResult = params.try_into().unwrap();
+        assert_eq!(result.name, "create-wiki-page");
         assert_eq!(result.path, "/My Page");
         assert_eq!(result.content, "Some wiki content here");
         assert!(result.comment.is_none());
@@ -413,14 +421,14 @@ mod tests {
 
     #[test]
     fn test_result_serializes_correctly() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Folder/Page".to_string(),
             content: "Sufficient content here".to_string(),
             comment: Some("initial commit".to_string()),
         };
-        let result: UpdateWikiPageResult = params.try_into().unwrap();
+        let result: CreateWikiPageResult = params.try_into().unwrap();
         let json = serde_json::to_string(&result).unwrap();
-        assert!(json.contains(r#""name":"update-wiki-page""#));
+        assert!(json.contains(r#""name":"create-wiki-page""#));
         assert!(json.contains(r#""path":"/Folder/Page""#));
     }
 
@@ -428,75 +436,97 @@ mod tests {
 
     #[test]
     fn test_validation_rejects_empty_path() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "".to_string(),
             content: "Some content here".to_string(),
             comment: None,
         };
-        let result: Result<UpdateWikiPageResult, _> = params.try_into();
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
         assert!(result.is_err());
     }
 
     #[test]
     fn test_validation_rejects_path_traversal() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/valid/../../../etc/passwd".to_string(),
             content: "Some content here".to_string(),
             comment: None,
         };
-        let result: Result<UpdateWikiPageResult, _> = params.try_into();
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
         assert!(result.is_err());
     }
 
     #[test]
     fn test_validation_rejects_null_bytes_in_path() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Page\x00Name".to_string(),
             content: "Some valid content here.".to_string(),
             comment: None,
         };
-        let result: Result<UpdateWikiPageResult, _> = params.try_into();
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
         assert!(result.is_err());
     }
 
     #[test]
     fn test_validation_rejects_short_content() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Page".to_string(),
             content: "short".to_string(),
             comment: None,
         };
-        let result: Result<UpdateWikiPageResult, _> = params.try_into();
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
         assert!(result.is_err());
     }
 
     #[test]
     fn test_validation_rejects_empty_content() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Page".to_string(),
             content: "".to_string(),
             comment: None,
         };
-        let result: Result<UpdateWikiPageResult, _> = params.try_into();
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
         assert!(result.is_err());
     }
 
     #[test]
     fn test_validation_accepts_valid_params() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Folder/My Page".to_string(),
             content: "This is sufficient content.".to_string(),
             comment: None,
         };
-        let result: Result<UpdateWikiPageResult, _> = params.try_into();
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
         assert!(result.is_ok());
+    }
+
+    #[test]
+    fn test_validation_rejects_bare_slash_path() {
+        let params = CreateWikiPageParams {
+            path: "/".to_string(),
+            content: "This is sufficient content.".to_string(),
+            comment: None,
+        };
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_validation_rejects_multiple_slash_only_path() {
+        let params = CreateWikiPageParams {
+            path: "///".to_string(),
+            content: "This is sufficient content.".to_string(),
+            comment: None,
+        };
+        let result: Result<CreateWikiPageResult, _> = params.try_into();
+        assert!(result.is_err());
     }
 
     // ── Config ────────────────────────────────────────────────────────────────
 
     #[test]
     fn test_config_defaults() {
-        let config = UpdateWikiPageConfig::default();
+        let config = CreateWikiPageConfig::default();
         assert!(config.wiki_name.is_none());
         assert!(config.wiki_project.is_none());
         assert!(config.branch.is_none());
@@ -512,15 +542,15 @@ wiki-name: "MyProject.wiki"
 wiki-project: "OtherProject"
 path-prefix: "/agent-output"
 title-prefix: "[Agent] "
-comment: "Updated by agent"
+comment: "Created by agent"
 "#;
-        let config: UpdateWikiPageConfig = serde_yaml::from_str(yaml).unwrap();
+        let config: CreateWikiPageConfig = serde_yaml::from_str(yaml).unwrap();
         assert_eq!(config.wiki_name.as_deref(), Some("MyProject.wiki"));
         assert_eq!(config.wiki_project.as_deref(), Some("OtherProject"));
         assert!(config.branch.is_none());
         assert_eq!(config.path_prefix.as_deref(), Some("/agent-output"));
         assert_eq!(config.title_prefix.as_deref(), Some("[Agent] "));
-        assert_eq!(config.comment.as_deref(), Some("Updated by agent"));
+        assert_eq!(config.comment.as_deref(), Some("Created by agent"));
     }
 
     #[test]
@@ -529,7 +559,7 @@ comment: "Updated by agent"
 wiki-name: "Azure Sphere"
 branch: "main"
 "#;
-        let config: UpdateWikiPageConfig = serde_yaml::from_str(yaml).unwrap();
+        let config: CreateWikiPageConfig = serde_yaml::from_str(yaml).unwrap();
         assert_eq!(config.wiki_name.as_deref(), Some("Azure Sphere"));
         assert_eq!(config.branch.as_deref(), Some("main"));
     }
@@ -539,7 +569,7 @@ branch: "main"
         let yaml = r#"
 wiki-name: "MyProject.wiki"
 "#;
-        let config: UpdateWikiPageConfig = serde_yaml::from_str(yaml).unwrap();
+        let config: CreateWikiPageConfig = serde_yaml::from_str(yaml).unwrap();
         assert_eq!(config.wiki_name.as_deref(), Some("MyProject.wiki"));
         assert!(config.path_prefix.is_none());
     }
@@ -592,24 +622,24 @@ wiki-name: "MyProject.wiki"
         // Use \x01 (SOH) — passes validate() but must be stripped by sanitize_fields().
         // Null bytes are rejected earlier at the validate() stage (see
         // test_validation_rejects_null_bytes_in_path).
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Page\x01Name".to_string(),
             content: "Some valid content here.".to_string(),
             comment: None,
         };
-        let mut result: UpdateWikiPageResult = params.try_into().unwrap();
+        let mut result: CreateWikiPageResult = params.try_into().unwrap();
         result.sanitize_fields();
         assert!(!result.path.contains('\x01'));
     }
 
     #[test]
     fn test_sanitize_preserves_path_structure() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Folder/My Page".to_string(),
             content: "Some valid content here.".to_string(),
             comment: None,
         };
-        let mut result: UpdateWikiPageResult = params.try_into().unwrap();
+        let mut result: CreateWikiPageResult = params.try_into().unwrap();
         result.sanitize_fields();
         assert_eq!(result.path, "/Folder/My Page");
     }
@@ -618,15 +648,15 @@ wiki-name: "MyProject.wiki"
 
     #[tokio::test]
     async fn test_execute_missing_wiki_name_returns_err() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Page".to_string(),
             content: "Some valid content here.".to_string(),
             comment: None,
         };
-        let mut result: UpdateWikiPageResult = params.try_into().unwrap();
+        let mut result: CreateWikiPageResult = params.try_into().unwrap();
         result.sanitize_fields();
 
-        let ctx = crate::tools::ExecutionContext {
+        let ctx = crate::safeoutputs::ExecutionContext {
             ado_org_url: Some("https://dev.azure.com/myorg".to_string()),
             ado_organization: Some("myorg".to_string()),
             ado_project: Some("MyProject".to_string()),
@@ -647,15 +677,15 @@ wiki-name: "MyProject.wiki"
 
     #[tokio::test]
     async fn test_execute_missing_org_url_returns_err() {
-        let params = UpdateWikiPageParams {
+        let params = CreateWikiPageParams {
             path: "/Page".to_string(),
             content: "Some valid content here.".to_string(),
             comment: None,
         };
-        let mut result: UpdateWikiPageResult = params.try_into().unwrap();
+        let mut result: CreateWikiPageResult = params.try_into().unwrap();
         result.sanitize_fields();
 
-        let ctx = crate::tools::ExecutionContext {
+        let ctx = crate::safeoutputs::ExecutionContext {
             ado_org_url: None,
             ..Default::default()
         };
@@ -676,20 +706,20 @@ wiki-name: "MyProject.wiki"
 
         let mut tool_configs = HashMap::new();
         tool_configs.insert(
-            "update-wiki-page".to_string(),
+            "create-wiki-page".to_string(),
             serde_json::json!({ "wiki-name": "Proj.wiki" }),
         );
 
         // Bypass validation by building the result directly (simulates a
         // tampered safe-output file that somehow smuggled ".." through).
-        let result = UpdateWikiPageResult {
-            name: "update-wiki-page".to_string(),
+        let result = CreateWikiPageResult {
+            name: "create-wiki-page".to_string(),
             path: "/valid/../etc/passwd".to_string(),
             content: "pwned".to_string(),
             comment: None,
         };
 
-        let ctx = crate::tools::ExecutionContext {
+        let ctx = crate::safeoutputs::ExecutionContext {
             ado_org_url: Some("https://dev.azure.com/myorg".to_string()),
             ado_organization: Some("myorg".to_string()),
             ado_project: Some("MyProject".to_string()),
@@ -713,21 +743,21 @@ wiki-name: "MyProject.wiki"
 
         let mut tool_configs = HashMap::new();
         tool_configs.insert(
-            "update-wiki-page".to_string(),
+            "create-wiki-page".to_string(),
             serde_json::json!({
                 "wiki-name": "Proj.wiki",
                 "path-prefix": "/agent-output"
             }),
         );
 
-        let result = UpdateWikiPageResult {
-            name: "update-wiki-page".to_string(),
+        let result = CreateWikiPageResult {
+            name: "create-wiki-page".to_string(),
             path: "/root-level-page".to_string(),
             content: "Some content here".to_string(),
             comment: None,
         };
 
-        let ctx = crate::tools::ExecutionContext {
+        let ctx = crate::safeoutputs::ExecutionContext {
             ado_org_url: Some("https://dev.azure.com/myorg".to_string()),
             ado_organization: Some("myorg".to_string()),
             ado_project: Some("MyProject".to_string()),
@@ -746,26 +776,26 @@ wiki-name: "MyProject.wiki"
     }
 
     #[tokio::test]
-    async fn test_execute_page_not_found_is_rejected() {
-        // When the page does not exist (404) the executor must fail —
-        // creation is not allowed by this tool.
+    async fn test_execute_page_already_exists_is_rejected() {
+        // When the page already exists the executor must fail —
+        // editing is not allowed by this tool.
         use std::collections::HashMap;
 
         let mut tool_configs = HashMap::new();
         tool_configs.insert(
-            "update-wiki-page".to_string(),
+            "create-wiki-page".to_string(),
             serde_json::json!({ "wiki-name": "Proj.wiki" }),
         );
 
         // Build result directly to bypass Stage-1 validation
-        let result = UpdateWikiPageResult {
-            name: "update-wiki-page".to_string(),
+        let result = CreateWikiPageResult {
+            name: "create-wiki-page".to_string(),
             path: "/Agent/Page".to_string(),
             content: "some content here".to_string(),
             comment: None,
         };
 
-        let ctx = crate::tools::ExecutionContext {
+        let ctx = crate::safeoutputs::ExecutionContext {
             ado_org_url: Some("https://dev.azure.com/myorg".to_string()),
             ado_organization: Some("myorg".to_string()),
             ado_project: Some("MyProject".to_string()),
@@ -780,24 +810,30 @@ wiki-name: "MyProject.wiki"
 
         // The GET will fail (network unreachable with a fake host), so the
         // executor returns an anyhow error. We only need to confirm the
-        // path-not-found guard is reachable; the no-network path verifies the
+        // already-exists guard is reachable; the no-network path verifies the
         // guard logic via the unit test below.
         let _ = result.execute_impl(&ctx).await;
         // (we cannot assert success/failure here without a real server;
-        //  the guard itself is exercised by test_page_not_found_guard_returns_failure)
+        //  the guard itself is exercised by test_page_already_exists_guard_returns_failure)
     }
 
-    /// Unit test for the page-not-found guard (no HTTP call needed).
+    /// Unit test for the page-already-exists guard logic.
+    ///
+    /// NOTE: This test verifies the conditional logic prototype in isolation —
+    /// it does *not* call `execute_impl` directly. If the guard were accidentally
+    /// removed from `execute_impl`, this test would still pass. The integration
+    /// tests in `tests/compiler_tests.rs` and the network-level test
+    /// `test_execute_page_already_exists_is_rejected` (which calls `execute_impl`
+    /// against a fake host) together catch regressions in the live code path.
     #[test]
-    fn test_page_not_found_guard_returns_failure() {
-        // Simulate the logic that replaced check_create_if_missing_guard:
-        // if !page_exists → failure.
-        let page_exists = false;
+    fn test_page_already_exists_guard_returns_failure() {
+        // Simulate the logic: if page_exists → failure.
+        let page_exists = true;
         let effective_path = "/Agent/Page";
-        let result = if !page_exists {
+        let result = if page_exists {
             Some(ExecutionResult::failure(format!(
-                "Wiki page '{effective_path}' does not exist. \
-                 Use a separate safe output to create new pages."
+                "Wiki page '{effective_path}' already exists. \
+                 Use the update-wiki-page safe output to update existing pages."
             )))
         } else {
             None
@@ -806,15 +842,18 @@ wiki-name: "MyProject.wiki"
         assert!(!result.unwrap().success);
     }
 
-    /// Confirm that an existing page (page_exists = true) proceeds past the guard.
+    /// Confirm that a non-existent page (page_exists = false) proceeds past the guard.
+    ///
+    /// NOTE: Same caveat as `test_page_already_exists_guard_returns_failure` above —
+    /// this tests the logic prototype, not the live `execute_impl` code path.
     #[test]
-    fn test_existing_page_passes_guard() {
-        let page_exists = true;
-        let effective_path = "/Agent/Page";
-        let result: Option<ExecutionResult> = if !page_exists {
+    fn test_new_page_passes_guard() {
+        let page_exists = false;
+        let effective_path = "/Agent/NewPage";
+        let result: Option<ExecutionResult> = if page_exists {
             Some(ExecutionResult::failure(format!(
-                "Wiki page '{effective_path}' does not exist. \
-                 Use a separate safe output to create new pages."
+                "Wiki page '{effective_path}' already exists. \
+                 Use the update-wiki-page safe output to update existing pages."
             )))
         } else {
             None
