+++ src/safeoutputs/missing_data.rs
@@ -4,7 +4,7 @@ use schemars::JsonSchema;
 use serde::Deserialize;
 
 use crate::tool_result;
-use crate::tools::Validate;
+use crate::safeoutputs::Validate;
 
 /// Parameters for reporting missing data
 #[derive(Deserialize, JsonSchema)]
@@ -37,7 +37,7 @@ tool_result! {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::ToolResult;
+    use crate::safeoutputs::ToolResult;
 
     #[test]
     fn test_result_has_correct_name() {
