+++ src/safeoutputs/noop.rs
@@ -2,7 +2,7 @@ use schemars::JsonSchema;
 use serde::Deserialize;
 
 use crate::tool_result;
-use crate::tools::Validate;
+use crate::safeoutputs::Validate;
 
 /// Parameters for describing a no operation. Use this if there is no work to do.
 #[derive(Deserialize, JsonSchema)]
@@ -27,7 +27,7 @@ tool_result! {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::ToolResult;
+    use crate::safeoutputs::ToolResult;
 
     #[test]
     fn test_result_has_correct_name() {
