+++ src/safeoutputs/reply_to_pr_comment.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 use super::PATH_SEGMENT;
 use crate::sanitize::{Sanitize, sanitize as sanitize_text};
 use crate::tool_result;
-use crate::tools::{ExecutionContext, ExecutionResult, Executor, Validate};
+use crate::safeoutputs::{ExecutionContext, ExecutionResult, Executor, Validate};
 use anyhow::{Context, ensure};
 
 /// Parameters for replying to an existing review comment thread on a pull request
@@ -240,7 +240,7 @@ impl Executor for ReplyToPrCommentResult {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::ToolResult;
+    use crate::safeoutputs::ToolResult;
 
     #[test]
     fn test_result_has_correct_name() {
