+++ src/safeoutputs/resolve_pr_thread.rs
@@ -9,7 +9,7 @@ use super::resolve_repo_name;
 use super::PATH_SEGMENT;
 use crate::sanitize::{Sanitize, sanitize as sanitize_text};
 use crate::tool_result;
-use crate::tools::{ExecutionContext, ExecutionResult, Executor, Validate};
+use crate::safeoutputs::{ExecutionContext, ExecutionResult, Executor, Validate};
 use anyhow::{Context, ensure};
 
 /// All valid thread status strings (lowercase, agent-facing)
@@ -286,7 +286,7 @@ impl Executor for ResolvePrThreadResult {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::ToolResult;
+    use crate::safeoutputs::ToolResult;
 
     #[test]
     fn test_result_has_correct_name() {
